dotCMS/core#11997 refactoring
