+++ src/app/portlets/content-types/create/content-types-create.component.html
@@ -6,7 +6,6 @@
         (onSubmit)="handleFormSubmit($event)"
         (onCancel)="router.navigate(['../'], { relativeTo: route })">
     </content-types-form>
-    <div *ngIf="readyToAddFields">
-        <content-type-fields-drop-zone></content-type-fields-drop-zone>
-    </div>
+
+    <content-type-fields-drop-zone></content-type-fields-drop-zone>
 </content-type-layout>
