+++ src/app/portlets/content-types/edit/content-types-edit.component.html
@@ -8,7 +8,8 @@
         (onSubmit)="handleFormSubmit($event)"
         (onCancel)="router.navigate(['../'], { relativeTo: route })">
     </content-types-form>
-    <div *ngIf="readyToAddFields">
-        <content-type-fields-drop-zone [fields]="data ? data.fields : []"></content-type-fields-drop-zone>
-    </div>
+
+    <content-type-fields-drop-zone *ngIf="data && data.fields"
+                                   [fields]="data.fields">
+                                </content-type-fields-drop-zone>
 </content-type-layout>
