+++ src/app/portlets/content-types/fields/common/field-column.ts
@@ -1,0 +1,5 @@
+import { Field } from './field';
+
+export interface FieldColumn {
+    fields: Field[];
+}
