+++ src/app/portlets/content-types/fields/common/field-type.ts
@@ -1,0 +1,3 @@
+export interface FieldType {
+    name: string;
+}
