+++ src/app/portlets/content-types/fields/common/field.ts
@@ -1,0 +1,13 @@
+
+export interface Field {
+    dataType: string;
+    fixed?: boolean;
+    indexed?: boolean;
+    listed?: boolean;
+    name?: string;
+    readOnly?: boolean;
+    required?: boolean;
+    sortOrder?: number;
+    unique?: boolean;
+    velocityVarName?: string;
+}
