+++ src/app/portlets/content-types/fields/common/index.ts
@@ -1,0 +1,4 @@
+export * from './field-column';
+export * from './field-row';
+export * from './field-type';
+export * from './field';
