+++ src/app/portlets/content-types/fields/content-type-fields-row-list/content-type-fields-row-list.component.html
@@ -1,6 +1,6 @@
-<ul  dragula="fields-row-bag"
-     [dragulaModel]="fieldRows"
-     data-drag-type="source">
+<ul dragula="fields-row-bag"
+    [dragulaModel]="fieldRows"
+    data-drag-type="source">
 
     <li *ngFor="let row of rows">Columns - {{row}}</li>
 </ul>
