+++ src/app/portlets/content-types/fields/content-types-fields-list/content-types-fields-list.component.html
@@ -1,6 +1,6 @@
-<ul  dragula="fields-bag"
-     [dragulaModel]="fieldTypes"
-     data-drag-type="source">
+<ul [dragulaModel]="fieldTypes"
+    dragula="fields-bag"   
+    data-drag-type="source">
 
     <li *ngFor="let fieldType of fieldTypes">{{fieldType.dataType}}</li>
 </ul>
