+++ src/app/portlets/content-types/fields/index.ts
@@ -3,3 +3,4 @@ export * from './content-type-fields-drop-zone';
 export * from './content-type-fields-row';
 export * from './content-type-field-dragabble-item';
 export * from './content-type-fields-row-list';
+export * from './common';
