+++ src/app/portlets/content-types/fields/service/field-drag-drop.service.ts
@@ -1,5 +1,5 @@
 
-import {Injectable} from '@angular/core';
+import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { DragulaService } from 'ng2-dragula';
 
@@ -15,29 +15,37 @@ export class FieldDragDropService {
 
     }
 
+    /**
+     * Set the options for the 'fields-bag' dragula group
+     * @memberof FieldDragDropService
+     */
     setFieldBagOptions(): void {
         let fieldBagOpts = this.dragulaService.find(FieldDragDropService.FIELD_BAG_NAME);
 
         if (!fieldBagOpts) {
             this.dragulaService.setOptions(FieldDragDropService.FIELD_BAG_NAME, {
                 copy: true,
-                moves: this.moves
+                moves: this.shouldMove
             });
         }
     }
 
+    /**
+     * Set the options for the 'fields-row-bag' dragula group
+     * @memberof FieldDragDropService
+     */
     setFieldRowBagOptions(): void {
         let fieldRowBagOpts = this.dragulaService.find(FieldDragDropService.FIELD_ROW_BAG_NAME);
 
         if (!fieldRowBagOpts) {
             this.dragulaService.setOptions(FieldDragDropService.FIELD_ROW_BAG_NAME, {
                 copy: true,
-                moves: this.moves
+                moves: this.shouldMove
             });
         }
     }
 
-    private moves(el, target): boolean {
+    private shouldMove(el, target): boolean {
         return target.dataset.dragType === 'source';
     }
 }
