+++ src/app/portlets/content-types/fields/service/field.service.ts
@@ -1,6 +1,6 @@
-import {Injectable} from '@angular/core';
+import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { Field, FieldType, FieldRow } from './';
+import { Field, FieldType, FieldRow } from '../';
 
 /**
  * Provide method to handle with the Field Types
