+++ src/app/portlets/content-types/form/content-types-form.component.scss
@@ -1,5 +1,5 @@
-@import "../../../../../styles/mixins";
-@import "../../../../../styles/variables";
+@import "../../../../styles/mixins";
+@import "../../../../styles/variables";
 
 .content-type__form-header {
     margin: $basic-padding * 3 0;
