+++ src/app/portlets/content-types/layout/content-types-layout.component.html
