+++ src/app/portlets/content-types/layout/content-types-layout.component.scss
@@ -1,5 +1,5 @@
-@import "../../../../../styles/mixins";
-@import "../../../../../styles/variables";
+@import "../../../../styles/mixins";
+@import "../../../../styles/variables";
 
 .content-type {
     .ui-tabview-nav {
