+++ src/app/portlets/content-types/layout/content-types-layout.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, ViewEncapsulation} from '@angular/core';
-import { BaseComponent } from '../../../../view/components/_common/_base/base-component';
-import { MessageService } from '../../../../api/services/messages-service';
+import { BaseComponent } from '../../../view/components/_common/_base/base-component';
+import { MessageService } from '../../../api/services/messages-service';
 import { DragulaService } from 'ng2-dragula';
 
 @Component({
