+++ src/app/portlets/content-types/layout/index.ts
